Fix loading multiple baselines from single git repository
This commit fixes somewhat rare case when one loads multiple different
baselines from single repository using `pharo-load-local` macro.

The problem was that `pharo-load-local` try to automagically register
IceBerg repository (as a convenience to the user). So when a second
baseline was loaded, Iceberg complained that given git repository is
already registered.

To fix that, this commit just ignores `IceDuplicatedRepository` thrown
when loading second (and subsequent) baseline.