Add nasty workaround to make Pharo work on systems with recent libcurl
On some Linux distros (e.g., Debian 12 Trixie) recent update of libcurl
broke Pharo. The problem is, that (Pharo-shipped) `libgit.so` links to
`libcurl-gnutls.so` which then links to `libssh2.so`. Pharo however ships
it's own `libssh2.so` which is incompatible with the one required by
`libcurl-gnutls.so`.

As of now, it seems that Pharo works with system `libssh2.so` too so
as a workaround, remove Pharo-shipped `libssh2.so` forcing Pharo to
use system one.

It is only a matter of a time when this or something else will stop
working. We'll see what to do when this happens.

Triple sigh.