Use custom Pharo 8 VM
This commit update makefiles to download our own build of Pharo 8 VM [1].
The main reason for this is that Pharo 8 VM no longer receives any
updates nor fixes (for example: [2], [3]). Another reason is to be
(to some degree) independent on third party services.

At this point, [1] only provides VMs for Linux x86_64. Windows and/or
MacOS VMs may be provided later, if needs be.

[1] https://github.com/janvrany/pharo-vm/releases/latest
[2] https://github.com/janvrany/smalltalk-makefiles/commit/59ea612a71761e0c46e91bf22b5cfdc3c8829c1f
[3] https://github.com/pharo-project/pharo-vm/issues/982